requirements.txt
fastapi
uvicorn
requests
gspread
google-auth
python-dotenv
openai
google-genai
Jinja2
gunicorn
